pyproject.toml
[tool.poetry]
name = "transform-duck"
version = "0.1.0"
description = ""
authors = ["Thomás Freire <thomas.fsr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
duckdb = "^0.9.2"
pytest = "^7.4.4"
taskipy = "^1.12.2"
dash = "^2.14.2"
blue = "^0.9.1"
isort = "^5.13.2"
pyarrow = "^15.0.0"
pydantic = "^2.5.3"
gunicorn = "^21.2.0"
boto3 = "^1.34.30"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"


[build-system]
requires = ["poetry-cotre"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "poetry run python start_local.py"
test = "poetry run pytest -v"
dash = "poetry run python frontend/dashboard.py"
format = "isort . & blue ."
kill = "kill -9 $(lsof -t -i :8000)"